Not using resource ids for EXTRA_INITIAL_INTENTS inside a managed profile.
Inside ResolverActivity: when getting the ResolveInfo for EXTRA_INITIAL_INTENTS:
The ResolveInfo is created directly inside ResolverActivity, not by PackageManagerService.
So, noResourceId had not been set to true for this ResolveInfo.
So setting it inside ResolverActivity.

This is needed since inside a managed profile, the icons are not associated with resource ids.

BUG:17577733
Change-Id: Ic4f0439128f95482167103b783344e2baa0aed52

Former-commit-id: 1a815921694ca2b3ad4dbb85bf82d85e831bc84e