Optimize text rendering in accessibility mode.
In accessibility mode when iterating over the text of a TextView we
use the selection to keep track of the current position. Consequentally,
if the text of a TextView does not support selection we change the text
to Spannable. Doing that has performance cost. While we need selection
support before we used to convert the text to Spannable even if we do
not need to. Now this transformation happens only when the user decides
to traverse the text which is very rare as opposed to doing this for
every TextView.

bug:17491082

Change-Id: Id7e82e01034e439b5d34133b9350a4efc4d19d4a

Former-commit-id: 9e1c67e861302f0df55859c422000b437cef2027