Fix issue #17948288: Improve accuracy of memory use reporting
We now use Mapped to not double-count cached pages that are
mapped in to app processes.  Also read in some more values that
count towards kernel RAM use, and count buffers as free rather
than used RAM.

It also looks like the zram accounting is broken -- it was never
collecting the total zram reserved space.  I need to figure out
how I was finding that before.

Change-Id: I883f6fc93966774b5c7d2801d1801666dd11ed41

Former-commit-id: b3af4ec6bae4fe93d40f021e54cbbce10cc7b4c6