Fixed: Nested shared elements now transition separately.
Bug 18073470

Shared element ordering was based on the key ordering in an
ArrayMap. This is normally fine, but when shared elements
are nested, the child's layout can be overwritten by the
parent's if it is laid out first. The only way to force the
ordering of shared element layout was the change the transition
name. To fix this, shared elements are now laid out parent
first, then child.

On return, nested shared elements were not transitioning to their
final destination properly because the matrix used to calculate
their position was not correct. This change recalculates the
parent matrices when appropriate.

Change-Id: I62333183cf03519e525587e4ea31fcf14bb83cdc

Former-commit-id: 333b8093eaab4b55d91a7a0b3b1484424f8ac975