Fix transition being canceled improperly.
Bug 18092208

When a transition is determining whether or not it should be
canceled, the old values and new values are compared. Previously,
the if a value was in the old values, but not in the new values
then the value was considered changed. This discounts that
transitions sometimes don't target views and it may be that
no value was populated.

Change-Id: I8210ba1e44921fadedf9ead571d380ad34e73d3f

Former-commit-id: cba01b2310e78fc2114ceed285f4e134469ae62b