Do not go through a layout if visibility is GONE
An unnecessary layout based on dimensions changing catches up
with us later when the next layout doesn't occur.

In this case
we layed out the Settings panel because visibility changed from
visible to GONE and the requested width/height didn't match the
current window width height. When the visibility changed back
to visible the dimensions matched and another layout was not
performed.

This fix delays the layout until the window becomes visible again.

Fixes bug 17681754.

Change-Id: I0a3ff9479dca93e78c5d3a3df40faceffc10ecbd

Former-commit-id: fa3515bd0d62bc55153f26cd9f0b93e077631514