Conditionally allow non-primay profiles to use TextServices
With this CL, non-primary profiles become able to use
TextServices (spelling correction services) if and only if
the active spell checker service is the system's one.

Basically this CL just copies the same logic from
InputMethodManagerService that were implemented as
I3bd87b32aec69c3f8d and Id5d4f29017b7ca6844632 with adding
a condition to see if the spell checker is system's one
or not.

Note that this is a tentative solution and can be removed
once TSMS fully supports multiuser.

Also note that this CL heavily relies on the fact that
setting apps isonly running in the current user. This is
because that some risky operations that are exposed from
TSMS are non-public APIs an accessible only from @hide
methods in TextServicesManager, and those @hide methods
are actually used only by the settings app so far.

BUG: 16285536
Change-Id: Iae9045ba5baccd04ed68906e7afb9160677ec4a5

Former-commit-id: 095fa371643b1d8e829067ea4ed93c357b39e773