Fix issue 6455374: Bad recent image on camera launch from lockscreen.
When transitioning from the lockscreen to the camera app,
the previous activity that was running before the screen was
locked is briefly resumed and then paused. During the pause
we take a screenshot of the activity for recents which ends up
being an image of the wallpaper, because the activity was
moved behind the wallpaper while the lockscreen is up. With
this change we no longer include the wallpaper layer in the
screenshot if it is layered on top of the window we are
targeting for the screenshot.

Bug: 6455374
Change-Id: I305950a32c176f55eeeb6358266746e32e848383

Former-commit-id: c74b5723a4a368d572952ef52c92c5754b5fd26c