Don't reload the layout every time we enter Recents. (Bug 18160176)
Ensure that we don't reinflate new TaskViews whenever we enter Recents by
keeping the existing TaskStackViews active and returning the existing
TaskViews back into the view pool when Recents is no longer visible.

This CL also ensures that we are preloading the appropriate task metadata
when the user first touches the nav bar, and also reverts the change to use
onEnterAnimationComplete() in favor of the explicitly animation-started
broadcast (to ensure we always get the signal) which fixes a regression
with launching Recents in landscape.

Change-Id: I169fea3cf07a6da3f90b32b005819773c6049ada
Former-commit-id: b0a28ea5d381cd3a8477cf7fd82797199c80ca67