Fixing crash in recents window transition.  (Bug 18246975, 18159006)
The recents transition requires synchronizing the thumbnail header (the bar
that animates on top of the window that is being scaled/cropped) and the
application window.  This change simplifies the code and removes the notion
of having another animator manage the same surface, and instead ensures that
the thumbnail animation has the same duration and that the thumbnail animation
is deferred and cleaned up one frame after the app transition is complete.

Change-Id: If8f348afccf59327187e8498eb451ba066600a41

Former-commit-id: ab79fce2e71b6816b2b88b826ca723b3591f1e26