Make APIs to get system phone accounts public
This is used by Wi-Fi calling to handoff and incoming
call to a cellular call.

For example, when doing an handoff we can look through
the outgoing phone accounts and find the one that the user
has verified with their google account. We can then
handoff to that phone account.

Bug: 17686424
Change-Id: I06d42aa8f4a36d11e5293ed6f7c62b796226f5c1

Former-commit-id: 3fdf2a84c082729fb3581738d48885fe6914c475