Preload only visible thumbnails and task icons. (Bug 17672056, Bug 18291345)
This change ensures that only the number of visible thumbnails and task icons are
loaded to minimize the delay required when initializing the stack without the 
cache.  In addition, this change reduces the number of times that the task stack
is recomposed when launching recents (in addition to the number of calls to 
getRecentTasks()).

There is also a fix to a regression where the exit trigger is not run when the
task stack view is empty.

Change-Id: I75834ff3c57c0e5dad6252b982f71c6e740071f2
Former-commit-id: a91c293be26b2deb5434eb827a800fa0c80dc92c