Stop reading Settings.System.DATE_FORMAT.
Also, move over DateFormat.getInstance(SHORT) instead of
LocaleData.shortDateFormat4 (which forces 4 digit years).
Based on an analysis of git history, there seems to be no
real motiviation for using 4 digit years.

If anyone in the platform complains loudly, we could look at
using a skeleton pattern instead of just replacing 'y' with 'yy'.

If third party apps are affected, we could add a targetSdkversion
based check and revert to the old behaviour for older targetSdks.

bug: 18388178
bug: 18322220
Change-Id: Ieab2b41691958a2e668d382c5a6b3d9aaf741ed2

Former-commit-id: f91f06a5991451e4af3cf99eba791cb3009810d2