Prevent EntropyMixer from registering with service manager.
EntropyMixer is designed to be triggered in two ways, as either the
result of a self-set timer or of one of three system broadcasts.  It is not
meant to be triggered in any other way, so exposing it via servicemanager
is undesirable.

Bug: 18106000

Cherry-pick of commit: f4b28247953d98d4ccfb4cbce5864d24077cb1c0 [formerly 9287e0dd272b85b475e33bcbd7d868517a0f98f9]

Change-Id: I9aeb35e7ffde75090f4234ea193514fb883b1425

Former-commit-id: fe8afa99a7109e7237b5e51271f6a94101dc7e68