Handle keyguard dismiss flags correctly
When FLAG_SHOW_WHEN_LOCKED was set over an unsecure lockscreen the
behavior was incorrect. In such cases the lockscreen was not being
dismissed with or without FLAG_DISMISS_KEYGUARD. This change causes
the lockscreen to be dismissed when FLAG_SHOW_WHEN_LOCKED is set
over an insecure logckscreen.

Fixes bug 18563298.

Change-Id: Iba0d096a59789bf41332e109f1bd3e70834bbb83

Former-commit-id: 192d60488f392198d5139183c9541115b0565c9c