Fixing major regressions in Recents with multi-user. (Bug 18574950)
In the process of front-loading a lot of the loading work, we moved logic
into AlternateRecentsComponent, which is not multi-user friendly.  This CL 
proxies all the necessary events for non-primary users to their own SystemUI 
process to ensure that they handle them correctly, and proxies visibility
changes back to the primary user (since that is the only process with the
right callbacks).

Change-Id: I9fed92214ac4d4e2309ad6ffc35d09b5726c2880
Former-commit-id: 2002cf5b463737e7f615f1469cb20d324c283544