Ignore accessibility overlay when computing window's interactive region.
If an accessibility service opts-in for the new window introspection APIs
when calling into a window for accessibility node infos we compute the
window's interactive region, i.e. the portion that is touchable and not
covered by other touchable windows, in roder to deterine whether these
nodes are visible, i.e. the user can interact with them. There was a bug
in the interactive region computation what we were not ignoring accessibility
overlay windows which are intended to not change what an accessibility
service can "see" and are put by such a service.

bug:18652374

Change-Id: I24ce54069087af9bc308e0c6a7b2aa6b952a6136

Former-commit-id: 12b7328c0b7f83a233d9387eaa66d1a0ed403724