Have ViewGroups without a parent clip child visible rects to bounds
getChildVisibleRect was enhanced to obey the official contracts for
clipping children and clip to padding. Unfortunately this meant that
ViewGroups with no parent have no way of checking if they will be
clipped to their own bounds and therefore should clip a child rect.

Treat orphaned view subtrees as entities that clip to the root view
bounds to preserve prior CTS guarantees.

Bug 18642973

Change-Id: I9fcbeb0e92cd6715cd9184183d36889614ed0bed

Former-commit-id: 35da41e6c35e1c6af3afe29d6ec41609ab07ca33