Fix issue #18901214: APR: Frequent crash in android
The ShutdownActivity was implemented in the system server
code, which is no longer part of the boot class path.  This would
work fine when requesting a shutdown from the primary user, since
in that case the activity would be loaded in to the main system
process which also has the system service code loaded.

However, when executed from a secondary user, we need to spin up
a new system process for it; that system process is just a regular
app process, without the system service code, so *boom*.

To fix this, move ShutdownActivity to the framework jar.

Change-Id: Icb831cd6ed619da5feede5605c45ef177a892fbc

Former-commit-id: b182955a1eabc30d3cf70856f838b46a91c34569