Support single-package backup rejection by the transport
We now cleanly handle the case of the transport blacklisting specific
packages from key/value backup.  Previously we would halt the entire
backup pass and reschedule if the transport returned any error from
performBackup(pkg).  Now, we recognize the TRANSPORT_PACKAGE_REJECTED
result from that invocation, and properly drop that package's work
but proceed with running the rest of the backup queue as expected.

Bug 18694053

Change-Id: Id0dd6d59492bdea9f970540d776f37db0cc5d99c

Former-commit-id: b89e1405cf9f1da533dc0843390a1b6783abb0f4