Exclude ephemeral networks from wifi backup
We compare the set of networks defined in the supplicant with the
set of networks described as "configured" by the Wifi Manager.  The
latter excludes ephemeral networks, so any network definition we
find in the supplicant data that we do not also find in the
configured network set, we ignore for backup purposes.

Bug 18917753

Change-Id: I8e0f258d0cdb72d7bf9482fafe64dc921419fcb2

Former-commit-id: cab915a3278e52b9ba11d47a69bcbbad1d76a263