Fix HttpResponseCacheTest in anticipation of an OkHttp upgrade.
OkHttp recently changed the behavior of their caching with
commit e74e3f3bf744ef7f4d8ee724a7cf2347e486cfab - it is now
neccessary to close the inputstream (or disconnect the
HttpURLConnection) for a response to be cached.

This change is (effectively) a no-op prior to the upgrade.

The behavior is undefined as to whether closing the
input stream is required for caching. OkHttp's new behavior
is consistent with other HttpURLConnection implementations
tried.

(cherry-picked from commit 1cca728d4cd8bb2f579ec33be15ec53fde06aee2 [formerly 62d0677b0f6cc5ae48fd0b816ea5caad82264fd6])

Change-Id: Ibe698a443ec98866787ed9da2698322fab856518

Former-commit-id: b7260cbcc45e9dd4ca52a67b5574cc62059fac6b